+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _Changes in the next release_
 
+### Added
+- Remote-entity support ([#23](https://github.com/unfoldedcircle/integration-home-assistant/issues/23)).
+
 ---
 
 ## v0.8.2 - 2024-03-04